+++ framework/PageDialogs/BulkActionDialog.tsx
@@ -10,11 +10,11 @@ import {
 import { CheckCircleIcon, ExclamationCircleIcon, PendingIcon } from '@patternfly/react-icons';
 import pLimit from 'p-limit';
 import { useCallback, useEffect, useState } from 'react';
-import { pfDanger, pfInfo, pfSuccess } from './components/pfcolors';
+import { pfDanger, pfInfo, pfSuccess } from '../components/pfcolors';
+import { ITableColumn, PageTable } from '../PageTable/PageTable';
+import { usePaged } from '../PageTable/useTableItems';
+import { useFrameworkTranslations } from '../useFrameworkTranslations';
 import { usePageDialog } from './PageDialog';
-import { ITableColumn, PageTable } from './PageTable/PageTable';
-import { usePaged } from './PageTable/useTableItems';
-import { useFrameworkTranslations } from './useFrameworkTranslations';
 
 export interface BulkActionDialogProps<T extends object> {
   /** The title of the model.
