+++ framework/PageDialogs/useSelectDialog.tsx
@@ -10,13 +10,16 @@ import {
   SplitItem,
 } from '@patternfly/react-core';
 import { useCallback, useEffect, useState } from 'react';
-import { Collapse } from './components/Collapse';
+import { Collapse } from '../components/Collapse';
+import { ITableColumn, PageTable, TableColumnCell } from '../PageTable/PageTable';
+import { IToolbarFilter } from '../PageTable/PageToolbar';
+import { ISelected } from '../PageTable/useTableItems';
+import { IView } from '../useView';
 import { usePageDialog } from './PageDialog';
-import { ITableColumn, PageTable, TableColumnCell } from './PageTable/PageTable';
-import { IToolbarFilter } from './PageTable/PageToolbar';
-import { ISelected } from './PageTable/useTableItems';
-import { IView } from './useView';
 
+/**
+ * @deprecated use SelectSingleDialog
+ */
 interface ISelectDialogOptions<T extends object, TMultiple> {
   view: IView & ISelected<T> & { itemCount?: number; pageItems: T[] | undefined };
   tableColumns: ITableColumn<T>[];
@@ -27,6 +30,9 @@ interface ISelectDialogOptions<T extends object, TMultiple> {
   isMultiple?: TMultiple extends true ? true : false;
 }
 
+/**
+ * @deprecated use SelectSingleDialog
+ */
 export function useSelectDialog<
   T extends { id: number; name: string | undefined },
   TMultiple = false
@@ -77,6 +83,9 @@ export function useSelectDialog<
   return openSetting;
 }
 
+/**
+ * @deprecated use SelectSingleDialog
+ */
 export type SelectDialogProps<T extends object, TMultiple> = {
   title: string;
   open: boolean;
