+++ framework/Settings.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable i18next/no-literal-string */
 import { Button, Form, Modal, ModalVariant, SelectOption } from '@patternfly/react-core';
 import { createContext, ReactNode, useCallback, useContext, useEffect, useState } from 'react';
-import { usePageDialog } from './PageDialog';
+import { usePageDialog } from './PageDialogs/PageDialog';
 import { FormGroupSelect } from './PageForm/Inputs/FormGroupSelect';
 import { useFrameworkTranslations } from './useFrameworkTranslations';
