+++ framework/index.ts
@@ -1,5 +1,3 @@
-export * from './BulkActionDialog';
-export * from './BulkConfirmationDialog';
 export * from './components/BulkSelector';
 export * from './components/Collapse';
 export * from './components/Help';
@@ -22,7 +20,11 @@ export * from './PageCells/TextCell';
 export * from './PageDetails/PageDetail';
 export * from './PageDetails/PageDetails';
 export * from './PageDetails/PageDetailsFromColumns';
-export * from './PageDialog';
+export * from './PageDialogs/BulkActionDialog';
+export * from './PageDialogs/BulkConfirmationDialog';
+export * from './PageDialogs/MultiSelectDialog';
+export * from './PageDialogs/PageDialog';
+export * from './PageDialogs/useSelectDialog';
 export * from './PageForm/Inputs/FormGroupSelect';
 export * from './PageForm/Inputs/FormGroupSelectOption';
 export * from './PageForm/Inputs/FormGroupTextArea';
@@ -47,6 +49,4 @@ export * from './PageTabs';
 export * from './Settings';
 export * from './useFrameworkTranslations';
 export * from './useInMemoryView';
-export * from './useSelectDialog';
-export * from './useSelectMultipleDialog';
 export * from './utils/compare';
