+++ frontend/awx/access/organizations/hooks/useAddOrganizationsToUsers.tsx
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useBulkActionDialog } from '../../../../../framework/BulkActionDialog';
+import { useBulkActionDialog } from '../../../../../framework/PageDialogs/BulkActionDialog';
 import { requestPost } from '../../../../Data';
 import { Organization } from '../../../interfaces/Organization';
 import { User } from '../../../interfaces/User';
